add pure cpp plate potential
had to make some other changes

there is now a harmonic interaction c++ potential

simple pairwise and neighborlist potentials now use the
add_energy_gradient functions so can be used in combined potential.

neighbor list potential now uses size_t instead of long int

+++ source/pele/wca.h
@@ -138,7 +138,7 @@ public:
  */
 class WCANeighborList : public SimplePairwiseNeighborList< WCA_interaction > {
 public:
-    WCANeighborList(Array<long int> & ilist, double sig, double eps)
+    WCANeighborList(Array<size_t> & ilist, double sig, double eps)
         :  SimplePairwiseNeighborList< WCA_interaction > (
                 std::make_shared<WCA_interaction>(sig, eps), ilist)
     {}
@@ -147,15 +147,15 @@ public:
 class WCAAtomList : public AtomListPotential<WCA_interaction, cartesian_distance<3>> {
 public:
     WCAAtomList(double sig, double eps, Array<size_t> atoms1, Array<size_t> atoms2)
-    : AtomListPotential<WCA_interaction, cartesian_distance<3> >(
+        : AtomListPotential<WCA_interaction, cartesian_distance<3> >(
             std::make_shared<WCA_interaction>(sig, eps),
             std::make_shared<cartesian_distance<3>>(), atoms1, atoms2)
-{}
+    {}
     WCAAtomList(double sig, double eps, Array<size_t> atoms1)
-    : AtomListPotential<WCA_interaction, cartesian_distance<3> >(
+        : AtomListPotential<WCA_interaction, cartesian_distance<3> >(
             std::make_shared<WCA_interaction>(sig, eps),
             std::make_shared<cartesian_distance<3>>(), atoms1)
-{}
+    {}
 };
 
 }
